requirements.txt
# Core dependencies
anthropic>=0.21.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector database
qdrant-client>=1.7.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Document parsing
pypdf>=4.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
tiktoken>=0.6.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Evaluation
scikit-learn>=1.4.0
